superstructures/ss1_gate/requirements.txt
flask
authlib
streamlit
boto3
python-dotenv
